1. 거리센서 연동 음성안내 작업액션 로직 개발

+++ WATA.LIS/WATA.LIS.Core/Model/VisionCam/VisionCamModel.cs
@@ -16,7 +16,7 @@ namespace WATA.LIS.Core.Model.VisionCam
         public string QR { get; set; }
         public int WIDTH { get; set; }
         public int HEIGHT { get; set; }
-        public int DEPTH { get; set; }
+        public float DEPTH { get; set; }
         public string POINTS { get; set; } // 형상측정값
         //public byte[] FRAME { get; set; }
         public bool connected = false;
